- moves comparers to a dedicated directory
