+++ src/Kiota.Builder/CodeDOM/CodeMethod.cs
@@ -2,6 +2,7 @@
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
+using Kiota.Builder.Comparers;
 using Kiota.Builder.Extensions;
 
 namespace Kiota.Builder.CodeDOM;
