+++ src/Kiota.Builder/CodeRenderers/CodeRenderer.cs
@@ -5,6 +5,7 @@ using System.Threading;
 using System.Threading.Tasks;
 
 using Kiota.Builder.CodeDOM;
+using Kiota.Builder.Comparers;
 using Kiota.Builder.Configuration;
 using Kiota.Builder.Writers;
