+++ src/Kiota.Builder/Comparers/OpenApiServerComparer.cs
@@ -4,7 +4,7 @@ using System.Diagnostics.CodeAnalysis;
 using System.Text.RegularExpressions;
 using Microsoft.OpenApi.Models;
 
-namespace Kiota.Builder;
+namespace Kiota.Builder.Comparers;
 
 internal sealed partial class OpenApiServerComparer : IEqualityComparer<OpenApiServer>
 {
