+++ src/Kiota.Builder/Writers/CSharp/CodeMethodWriter.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 
 using Kiota.Builder.CodeDOM;
+using Kiota.Builder.Comparers;
 using Kiota.Builder.Extensions;
 
 namespace Kiota.Builder.Writers.CSharp;
