+++ src/Kiota.Builder/Writers/Go/GoCodeParameterOrderComparer.cs
@@ -1,4 +1,5 @@
 ﻿using Kiota.Builder.CodeDOM;
+using Kiota.Builder.Comparers;
 
 namespace Kiota.Builder.Writers.Go;
