+++ src/Kiota.Builder/Writers/Java/CodeMethodWriter.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text.RegularExpressions;
 using Kiota.Builder.CodeDOM;
+using Kiota.Builder.Comparers;
 using Kiota.Builder.Extensions;
 
 namespace Kiota.Builder.Writers.Java;
