+++ src/Kiota.Builder/Writers/Php/CodeMethodWriter.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using Kiota.Builder.CodeDOM;
+using Kiota.Builder.Comparers;
 using Kiota.Builder.Extensions;
 namespace Kiota.Builder.Writers.Php;
 public class CodeMethodWriter : BaseElementWriter<CodeMethod, PhpConventionService>
