+++ src/Kiota.Builder/Writers/Python/PythonCodeParameterOrderComparer.cs
@@ -1,4 +1,5 @@
 ﻿using Kiota.Builder.CodeDOM;
+using Kiota.Builder.Comparers;
 
 namespace Kiota.Builder.Writers.Python;
 public class PythonCodeParameterOrderComparer : BaseCodeParameterOrderComparer
